+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fix sorting tasks when "Group tasks" is enabled
 * Fix the bug with task labels background opacity resetting to 100 when specifying 0 opacity
 * Remove interactivity-utils
+* Apply font settings to "Collapse All/Expand All" button
 
 
 # 3.0.13